Add Swagger API documentation in Account microservice
Included Swagger annotations in the Controllers and DTO's in the 'accounts' microservice to generate API documentation. Swagger helps to describe the structure of APIs in a manner that is both human and machine-readable which is crucial for testing, building, and managing APIs. Added springdoc-openapi-ui dependency to the pom.xml to implement Swagger-UI, an interactive API documentation where API endpoints can be tested directly in the browser.