Refactor docker-compose files and update application versions
The existing docker-compose file has been removed and replaced with new docker-compose files that better manage the common configurations. The image version has been upgraded from v1.0.0 to v1.1.0 in the cards, accounts, loans, and configserver services to use the latest versions of the applications. Additionally, readiness and liveness probes are now enabled in the config server.