Update image versions and add eureka server in docker-compose
Updated the version of docker images in the pom.xml and docker-compose files from v1.2.0 to v1.3.0 for improved functionalities. Added a new Eureka server configuration in docker-compose files to enable service registration and discoverability. Also updated dependencies and services configuration as part of the integration of the Eureka server.