Add message service with communication functionalities
Added new components to the messaging microservice for communicating with account services. These components include DTO classes, Docker files, and configurations for email and SMS functions. The '.gitignore' file was also added to manage unrelated files. An application.yml was created for the messaging service with specific routings for rabbitmq and specification for the server port. Additionally, Maven support was set up with relevant scripts and configurations.